git-ref-format: Fix dependence on re-export
The `refname!` macro emits code that only complies if a `RefString` can
be imported via the re-export in `radicle_git_ext`. This does not cause
any problems as long as `radicle_git_ext` is accessible during
compilation, which will be the case very often. However, it should also
be possible to use `git-ref-format`, and the `refname!` macro it
defines, without also depending on `radicle-git-ext`.

This is specifically important because of the very different dependency
closure of `git-ref-format` vs. `radicle-git-ext`: The former only has a
small dependency footprint (the only non optional dependency is
`thiserror` and of course its closure) while the latter depends on
`git2`, thus pulls in a whole implementation of Git!

The change itself is simple enough, it just changes the import to refer
to `git_ref_format` directly instead of going via `radicle-git-ext`.